+++ python/nano/src/bigdl/nano/deps/openvino/core/model.py
@@ -22,7 +22,7 @@ from openvino.tools.pot.graph import load_model, save_model
 from openvino.tools.pot.engines.ie_engine import IEEngine
 from openvino.tools.pot.pipeline.initializer import create_pipeline
 from openvino.tools.pot.graph.model_utils import compress_model_weights
-from .openvino_utils import save
+from .utils import save
 
 
 class OpenVINOModel:
